requirements.txt
langchain
langchain-community
langchain-ollama
langgraph
chromadb
python-dotenv
pydantic
httpx
pypdf